requirements.txt
addict==2.2.1
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.2
django-compat==1.0.15
django-cors-headers==2.5.2
django-rest-swagger==2.2.0
django-stubs==0.12.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
Faker==1.0.5
future==0.17.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
kavenegar==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
mypy==0.701
mypy-extensions==0.4.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2-binary==2.8.3
pyaes==1.6.1
pycparser==2.19
PyJWT==1.7.1
Pyrogram==0.8.0
pyrsistent==0.15.4
PySocks==1.6.8
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
StringGenerator==0.3.3
swagger-spec-validator==2.4.3
text-unidecode==1.2
TgCrypto==1.1.1
typed-ast==1.3.5
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
